requirements.txt
Flask==3.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
bcr-api
